requirements.txt
altair
certifi
Flask
Jinja2
joblib
jupyter
numpy
MonsterLab
pandas
pymongo[srv]
python-dotenv
scikit-learn
scipy
uvicorn
